% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met2Graph.R
\name{MetExtract}
\alias{MetExtract}
\title{MetExtract.R}
\usage{
MetExtract(infile)
}
\arguments{
\item{infile}{Path to the input SBML file}
}
\value{
model and data frames of reactants and products for each reaction. The output values are stored in a list.
}
\description{
This function loads a metabolic model in SBML format and returns a list object containing the model and data frames of reactants and products for each reaction.
}
\examples{
\dontrun{
out <- MetExtract(system.file("extdata", "kidney.xml", package = "Met2Graph"))}
}
