% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Met2Graph.R
\name{Plot_graphs}
\alias{Plot_graphs}
\title{Plot_graphs}
\usage{
Plot_graphs(inDir, outDir, format, ...)
}
\arguments{
\item{inDir}{path to the input directory.}

\item{outDir}{path to the output directory.}

\item{format}{Character string indicating the format of the graph to read. Default "ncol". For other formats and details please see read.graph function of igraph package.}

\item{...}{Additional plotting parameters. See igraph.plotting for the complete list.}
}
\value{
A pdf file.
}
\description{
This function plots graphs to pdf.
}
\examples{
\dontrun{
inDir <- system.file("extdata/output/MetGraphs/", package = "Met2Graph")
outDir <- system.file("extdata/output/", package = "Met2Graph")
Plot_graphs(inDir=inDir, outDir=outDir, format="ncol", vertex.size= 1,edge.arrow.size=0.1,vertex.label=NA, layout=layout_with_graphopt)}
}
